Let transport buttons have their mode toggled
This introduces an unfortunate name confusion:
* "recordMode", which sets the button mode for the Control Groups (either one, or all of them), vs
* "recordButtonMode", which sets the button mode for the Transport Button

Idk how to best resolve this yet. For now, I'll just leave it, and stay
away from SemVer 1.0, so I can freely undo this mistake. (Nevermind that
library.properties lists this as 1.0 already. Ugh.)

Also, NB: the track and marker buttons don't have methods to control their
mode! That's because the nanokontrol2 doesn't light up their button LEDs
(afaict) when they're pressed. That's a PITA in real usage.

The textfile I've been using for spelunking has this table:

    4.Native KORG Mode Messages
    ----------------------------------------------------

    (1) Native KORG mode Display LEDs
    +--------+------------------+--------------------------------------+
    | Status | Second   | Third |   Description                        |
    | [Hex]  | [Hex]    | [Hex] |                                      |
    +--------+----------+-------+--------------------------------------+
    |   BF   |  2E      |  ss   |   Cycle (ss= Off:00~40 On:41~7F)     |
    |   BF   |  2B      |  ss   |   REW                                |
    |   BF   |  2C      |  ss   |   FF                                 |
    |   BF   |  2A      |  ss   |   STOP                               |
    |   BF   |  29      |  ss   |   PLAY                               |
    |   BF   |  2D      |  ss   |   REC                                |
    +--------+----------+-------+--------------------------------------+
    |   BF   |  20      |  ss   |   Group 1 Solo                       |
    |   BF   |  21 ~ 27 |  ss   |   Group 2 ~ 8 Solo                   |
    |   BF   |  30      |  ss   |   Group 1 Mute                       |
    |   BF   |  31 ~ 37 |  ss   |   Group 2 ~ 8 Mute                   |
    |   BF   |  40      |  ss   |   Group 1 Rec                        |
    |   BF   |  41 ~ 47 |  ss   |   Group 2 ~ 8 Rec                    |
    +--------+----------+-------+--------------------------------------+

Notice that it doesn't list the track and marker buttons. Korg seems to
expect these to never be toggleable (though they do appear in the scene
data, and you *can* set them by passing the right scene data back to the
device).

If I change my mind, all the workings are in place - I just have to add
more methods like Kontrol2#cycleButtonMode; all the Scene bits are
already in place. If I do that, it might be nice to add a warning:
"You're setting these to Toggle, and it'll work, but FYI, the LEDs won't
light up. It's not a bug, it's the way it's implemented."

Finally, also NB: setting the button mode for transport buttons seems to
persist across launching a sketch. I don't know why, since the control
group button modes seem to reset.